refactor: :rocket: We adjusted details for the notifications section

+++ Bohemia/src/main/java/com/redsocial/bohemia/domain/service/NotificationServiceImpl.java
@@ -4,17 +4,19 @@ import com.redsocial.bohemia.domain.repository.NotificationRepository;
 import com.redsocial.bohemia.domain.repository.UserRepository;
 import com.redsocial.bohemia.persistence.entity.Notification;
 import com.redsocial.bohemia.persistence.entity.User;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
 import java.util.Date;
 import java.util.List;
 import java.util.Optional;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
 @Service
 public class NotificationServiceImpl implements NotificationService {
 
     @Autowired
     private NotificationRepository notificationRepository;
+
     @Autowired
     private UserRepository userRepository;
 
@@ -25,7 +27,11 @@ public class NotificationServiceImpl implements NotificationService {
 
     @Override
     public void delNotification(Long id_notification) {
-        notificationRepository.deleteById(id_notification);
+        if (notificationRepository.existsById(id_notification)) {
+            notificationRepository.deleteById(id_notification);
+        } else {
+            throw new RuntimeException("Notificación no encontrada");
+        }
     }
 
     @Override
@@ -36,11 +42,11 @@ public class NotificationServiceImpl implements NotificationService {
     @Override
     public Optional<Notification> findNotification(Long id_notification) {
         return notificationRepository.findById(id_notification);
-
     }
 
     public void notifyUser(Long userId, String message) {
-        User user = userRepository.findById(userId).orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
         Notification notification = new Notification(message, user, new Date());
         notificationRepository.save(notification);
     }
@@ -50,5 +56,4 @@ public class NotificationServiceImpl implements NotificationService {
                 .orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
         return notificationRepository.findByUserAndReadFalse(user);
     }
-
 }
